HIA-1635 Refactor redactions package
