+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/redaction/policies/PrisonEducation.kt
@@ -1,8 +1,8 @@
-package uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction
+package uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.policies
 
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.redactionconfig.RedactionType.MASK
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.redactionconfig.RedactionType.REMOVE
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.dsl.redactionPolicy
+import uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.redactionPolicy
 
 val prisonEducationRedactionPolicy =
   redactionPolicy("prison-education") {
