+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/redaction/policies/riskScores/EpfRedactions.kt
@@ -1,7 +1,7 @@
-package uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.riskScores
+package uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.policies.riskScores
 
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.redactionconfig.RedactionType.REMOVE
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.dsl.redactionPolicy
+import uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.redactionPolicy
 
 val epfRiskScoreRedactions =
   redactionPolicy(
