+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/roles/FullAccessLaoRedactions.kt
@@ -1,8 +1,8 @@
 package uk.gov.justice.digital.hmpps.hmppsintegrationapi.roles
 
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.roleconfig.roleConstants
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.laoRedactionPolicy
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.riskScores.generalRiskScoreRedactions
+import uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.policies.laoRedactionPolicy
+import uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.policies.riskScores.generalRiskScoreRedactions
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.roles.dsl.role
 
 val fullAccessLaoRedactions =
