+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/roles/Mappa.kt
@@ -1,6 +1,6 @@
 package uk.gov.justice.digital.hmpps.hmppsintegrationapi.roles
 
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.riskScores.generalRiskScoreRedactions
+import uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.policies.riskScores.generalRiskScoreRedactions
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.roles.dsl.role
 
 val mappa =
