+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/roles/TestRoleIdOnly.kt
@@ -1,6 +1,6 @@
 package uk.gov.justice.digital.hmpps.hmppsintegrationapi.roles
 
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.dsl.personSearchIdOnly
+import uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.policies.personSearchIdOnly
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.roles.dsl.role
 
 val testRoleWithIdOnlyRedaction =
